Move PopoverProps from Popover component to its own module
